+++ src/app/(board)/board/[boardId]/_components/board-detail/index.tsx
@@ -2,15 +2,13 @@
 
 import { useQuery } from "@tanstack/react-query";
 import clsx from "clsx";
-import { doc, getDoc, setDoc, updateDoc } from "firebase/firestore";
 import { useAtom } from "jotai";
 import Image from "next/image";
 import { redirect, useRouter } from "next/navigation";
-import { useEffect, useLayoutEffect, useMemo, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 
 import { Button } from "@/components/common";
 import EMPTY_IMAGE from "@/constants/image";
-import FIREBASE_DB from "@/firebase";
 import getBoardDetailData from "@/lib/api/board/get-board-detail-data";
 import { IconComment } from "@/public/assets/icons";
 import { userAtom } from "@/stores";
@@ -23,12 +21,12 @@ import CopyTeamToken from "./copy-team-token";
 interface BoardDetailProps {
   userId: number;
   boardId: number;
+  viewCount: number;
 }
 
-const BoardDetail = ({ userId, boardId }: BoardDetailProps) => {
+const BoardDetail = ({ userId, boardId, viewCount }: BoardDetailProps) => {
   const router = useRouter();
   const [previousPage, setPreviousPage] = useState<string | null>("");
-  const [viewCount, setViewCount] = useState<number | null>(0);
 
   const [user] = useAtom(userAtom);
 
@@ -45,29 +43,6 @@ const BoardDetail = ({ userId, boardId }: BoardDetailProps) => {
 
     setPreviousPage(previousUrl);
   }, []);
-
-  useLayoutEffect(() => {
-    const getViewCount = async () => {
-      const docRef = doc(FIREBASE_DB, "boards", boardId.toString());
-
-      // 조회수 가져오기
-      const boardDb = await getDoc(docRef);
-      const boardData = boardDb.data();
-
-      if (boardData) {
-        // 조회수 업데이트
-        const currentViewCount = boardData.viewCount + 1;
-        setViewCount(currentViewCount);
-        updateDoc(docRef, { viewCount: currentViewCount });
-      } else {
-        // 조회수 생성
-        await setDoc(docRef, { viewCount: 1 });
-        setViewCount(1);
-      }
-    };
-
-    getViewCount();
-  }, [boardId]);
 
   const { data: boardData, error } = useQuery({
     queryKey: ["board", boardId],
@@ -150,7 +125,7 @@ const BoardDetail = ({ userId, boardId }: BoardDetailProps) => {
             height={0}
             sizes="100vw"
             onDragStart={(e) => e.preventDefault()}
-            className="h-auto w-full"
+            className="h-auto min-h-200 min-w-327"
           />
         </div>
       )}
