+++ src/app/(board)/board/[boardId]/page.tsx
@@ -3,9 +3,11 @@ import {
   HydrationBoundary,
   QueryClient,
 } from "@tanstack/react-query";
+import { doc, getDoc, setDoc, updateDoc } from "firebase/firestore";
 import { Metadata } from "next";
 import { cookies } from "next/headers";
 
+import FIREBASE_DB from "@/firebase";
 import getBoardDetailData from "@/lib/api/board/get-board-detail-data";
 import getBoardComment from "@/lib/api/board-comment/get-comment";
 
@@ -34,6 +36,20 @@ const BoardPage = async ({ params }: { params: { boardId: number } }) => {
 
   const { boardId } = params;
   const userId = cookies().get("userId")?.value;
+  let viewCount = 1;
+
+  const docRef = doc(FIREBASE_DB, "boards", boardId.toString());
+
+  const boardDb = await getDoc(docRef);
+  const boardData = boardDb.data();
+
+  if (boardData) {
+    const currentViewCount = boardData.viewCount + 1;
+    viewCount = currentViewCount;
+    updateDoc(docRef, { viewCount: currentViewCount });
+  } else {
+    await setDoc(docRef, { viewCount: 1 });
+  }
 
   await Promise.all([
     queryClient.fetchQuery({
@@ -49,7 +65,11 @@ const BoardPage = async ({ params }: { params: { boardId: number } }) => {
 
   return (
     <HydrationBoundary state={dehydrate(queryClient)}>
-      <BoardDetail userId={Number(userId)} boardId={boardId} />
+      <BoardDetail
+        viewCount={viewCount}
+        userId={Number(userId)}
+        boardId={boardId}
+      />
       <CommentList boardId={boardId} />
     </HydrationBoundary>
   );
